Fixing a small typo in the previous commit
git-svn-id: http://libopenmetaverse.googlecode.com/svn/libopenmetaverse/trunk@2491 52acb1d6-8a22-11de-b505-999d5b087335

+++ Programs/Simian/LLUDP/LLMap.cs
@@ -296,7 +296,7 @@ namespace Simian
                                 Logger.Log(String.Format("HyperGrid teleporting to {0} ({1}, {2}) @ {3}",
                                     link.RegionName, x, y, destination), Helpers.LogLevel.Info);
 
-                                OSDMap teleport = CapsMessages.TeleportFinish(agent.ID, 4, link.RegionHandle, seedCap, (byte)SimAccess.Min,
+                                OSDMap teleport = CapsMessages.TeleportFinish(agent.ID, 4, link.RegionHandle, seedCap, SimAccess.Min,
                                     entry.AddressList[0], link.UDPPort, TeleportFlags.ViaLocation);
 
                                 scene.SendEvent(agent, "TeleportFinish", teleport);
